Split different test Add some tests for utility and app
